pyproject.toml
[tool.poetry]
name = "juggler"
version = "0.1.0"
description = ""
authors = ["Thiago Cardoso <thiagoncc@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
textual = "^0.53.1"
pyyaml = "^6.0.1"
litellm = "^1.40.25"
pydantic = "^2.10.4"


[tool.poetry.group.dev.dependencies]
pyright = "^1.1.391"

[tool.pyright]
include = ["juggler"]
#strict = ["juggler"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
